Move omps_util.py into utils/
*OSBS-8605

Also change files which import omps_util.py
